+++ src/svlog/lib.rs
@@ -40,12 +40,6 @@ extern crate log;
 pub extern crate moore_svlog_syntax as syntax;
 pub(crate) use moore_common as common;
 
-// Inline the salsa crate as a module, since we use a very esoteric branch for
-// now.
-// TODO(fschuiki): Remove this once salsa is regular dep again
-#[macro_use]
-mod salsa;
-
 /// Assert that a condition holds, or emit a bug diagnostic and panic.
 #[macro_export]
 macro_rules! assert_span {